Test suite audit cleanup (#1219)
- Delete dead rust/http-gateway/src/config.rs (unused, duplicated
  public::servers::http_gateway::HeaderForwardingConfig and its tests)
  and update mkdocs gateway docs to point at the real implementation
- Move inline #[cfg(test)] blocks out of tracing/src/time.rs,
  tracing/src/logs/events.rs, and tracing/src/string_id.rs into
  tests/time_tests.rs, tests/log_events_tests.rs, and
  tests/string_id_tests.rs, per the tests-under-tests/ convention
- Mark ingestion's check_ready_returns_true_when_dependencies_healthy
  as #[ignore] instead of silently no-oping when
  MICROMEGAS_SQL_CONNECTION_STRING is unset, matching the analytics
  crate's convention for live-dependency tests
- Replace the fixed 50ms sleep in large_message_tests.rs with a
  bounded TCP connect-retry poll so the test doesn't guess how long
  the mock server takes to start accepting connections
