+++ rust/tracing/src/string_id.rs
@@ -60,23 +60,3 @@ impl StringId {
         self.ptr as u64
     }
 }
-
-#[cfg(test)]
-mod test {
-    #[test]
-    fn test_string_id() {
-        use super::*;
-        let string_id = StringId::from("hello");
-        assert_eq!(string_id.len, 5);
-        assert_eq!(string_id.ptr, "hello".as_ptr());
-        assert_eq!(string_id.id(), "hello".as_ptr() as u64);
-
-        let mut buffer = vec![];
-        string_id.write_value(&mut buffer);
-        assert_eq!(buffer.len(), std::mem::size_of::<StringId>());
-
-        let string_id = unsafe { StringId::read_value(&buffer) };
-        assert_eq!(string_id.len, 5);
-        assert_eq!(string_id.ptr, "hello".as_ptr());
-    }
-}
