+++ rust/tracing/tests/time_tests.rs
@@ -1,0 +1,4 @@
+#[test]
+fn test_frequency() {
+    eprintln!("cpu frequency: {}", micromegas_tracing::time::frequency());
+}
